requirements.txt
discord.py==1.5.1
python-dotenv==0.15.0
google==3.0.0